+++ changelog.md
@@ -1,3 +1,16 @@
+# 3.18.0
+
+Updated for patch 11.2.5.
+
+## Changes
+
+- Trade time remaining checks are no longer performed in combat as they would be easy to overlook.
+
+## Bugfixes
+
+- *Fixed issue with running `/rc add` commands after having declined usage could lead to errors.*
+- *Leaving a group will now properly disable the voting frame fixing certain errors that could happen with it being enabled.*
+
 # 3.17.6
 
 ## Bugfixes